% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_force_data.R
\name{as_force_data}
\alias{as_force_data}
\title{Convert one of numerous data types to force_network's 'native' data format}
\usage{
as_force_data(.data, ...)
}
\arguments{
\item{.data}{a force network description in one of numerous forms (see
details).}

\item{...}{other arguments that will be passed on to as_force_data}
}
\description{
The \code{force_network} function uses a 'native' data format that consists of a...
}
