% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sankey_data.R
\name{as_sankey_data}
\alias{as_sankey_data}
\title{Convert one of numerous data types to sankey_network's 'native' data format}
\usage{
as_sankey_data(.data, ...)
}
\arguments{
\item{.data}{a sankey network description in one of numerous forms (see
details).}

\item{...}{other arguments that will be passed on to as_sankey_data}
}
\description{
The \code{sankey_network} function uses a 'native' data format that consists of a...
}
