% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tree_data.R
\name{as_tree_data.igraph}
\alias{as_tree_data.igraph}
\title{Convert igraph tree to \code{treenetdf}}
\usage{
\method{as_tree_data}{igraph}(data, root = "root", ...)
}
\arguments{
\item{data}{a tree network description in one of numerous forms (see
details).}

\item{root}{character specifying the string that should be used to name the
root node}

\item{...}{arguments to pass to methods.}
}
\description{
Convert igraph tree to \code{treenetdf}
}
