% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tree_data.R
\name{as_tree_data.Node}
\alias{as_tree_data.Node}
\title{data.tree to \code{treenetdf}}
\usage{
\method{as_tree_data}{Node}(data, ...)
}
\arguments{
\item{data}{a tree network description in one of numerous forms (see
details).}

\item{...}{arguments to pass to methods.}
}
\description{
data.tree to \code{treenetdf}
}
