% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tree_data.R
\name{as_tree_data}
\alias{as_tree_data}
\alias{as_tree_data.character}
\alias{as_tree_data.hclust}
\alias{as_tree_data.list}
\alias{as_tree_data.Node}
\alias{as_tree_data.phylo}
\alias{as_tree_data.tbl_graph}
\alias{as_tree_data.igraph}
\alias{as_tree_data.data.frame}
\title{Convert one of numerous data types to tree_network's 'native' treenetdf form}
\usage{
as_tree_data(data, ...)

\method{as_tree_data}{character}(data, ...)

\method{as_tree_data}{hclust}(data, ...)

\method{as_tree_data}{list}(data, children_name = "children", node_name = "name", ...)

\method{as_tree_data}{Node}(data, ...)

\method{as_tree_data}{phylo}(data, ...)

\method{as_tree_data}{tbl_graph}(data, ...)

\method{as_tree_data}{igraph}(data, root = "root", ...)

\method{as_tree_data}{data.frame}(
  data,
  cols = NULL,
  df_type = "treenetdf",
  subset = names(data),
  root,
  ...
)
}
\arguments{
\item{data}{a tree network description in one of numerous forms (see
details).}

\item{...}{other arguments that will be passed on to as_tree_data}

\item{children_name}{character specifying the name used for the list element
that contains the childeren elements.}

\item{node_name}{character specifying the name used for the list element that
contains the name of the node}

\item{root}{root name.}

\item{cols}{named character vector specifying the names of columns to be
converted to the standard \code{treenetdf} names.}

\item{df_type}{character specifying which type of data frame to convert. Can
be \code{treenetdf} or
\code{leafpathdf}.}

\item{subset}{character vector specifying the names of the columns (in order)
that should be used to define the hierarchy.}
}
\description{
The \code{tree_network} function uses a 'native' data format that consists of a data
frame with minimally 2 vectors/columns, one named \code{'nodeId'} and one named
\code{'parentId'}. Other columns in the data frame are also passed on to the
JavaScript code and attached to the elements in the D3 visualization so that
they can potentially be accessed by other JavaScript functions. This is an
advantageous format because:
\itemize{
\item it's an easy to use and understand R-like format
\item a hierarchical network can be succinctly defined by a list of each unique
node and its parent node
\item since each row defines a unique node, additional columns can be added to
add node-specific properties
\item in a hierarchical network, every link/edge can be uniquely identified by
the node which it leads to, therefore each link/edge can also be specifically
addressed by adding columns for formatting of the incoming link
}

\code{as_tree_data} can convert from any of the following data types:
\itemize{
\item \code{leafpathdf} (table)--\code{parent|parent|node}--\code{data.frame}
\item hierarchical nested list (JSON)
\item \code{hclust}
\item \code{data.tree} Node
\item igraph
\item ape \code{phylo}
}
}
\section{Methods (by class)}{
\itemize{
\item \code{as_tree_data(character)}: Convert JSON from URL to \code{treenetdf}

\item \code{as_tree_data(hclust)}: Convert hclust objects to \code{treenetdf}

\item \code{as_tree_data(list)}: Convert a nested list to \code{treenetdf}

\item \code{as_tree_data(Node)}: data.tree to \code{treenetdf}

\item \code{as_tree_data(phylo)}: Phylo tree to \code{treenetdf}

\item \code{as_tree_data(tbl_graph)}: tbl_graph_to_treenetdf

\item \code{as_tree_data(igraph)}: Convert igraph tree to \code{treenetdf}

\item \code{as_tree_data(data.frame)}: Convert a data.frame to a \code{treenetdf}

}}
\examples{
links <- read.csv(header = TRUE, stringsAsFactors = FALSE, text = '
                   source,target,name
                   1,,one
                   2,1,two
                   3,1,three
                   4,1,four
                   5,2,five
                   6,2,six
                   7,2,seven
                   8,6,eight')

 # Convert data
 as_tree_data(links, cols = c(nodeId = 'source', parentId = 'target'))

}
