% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tree_data.R
\name{as_tree_data.tbl_graph}
\alias{as_tree_data.tbl_graph}
\title{tbl_graph_to_treenetdf}
\usage{
\method{as_tree_data}{tbl_graph}(data, ...)
}
\arguments{
\item{data}{a tree network description in one of numerous forms (see
details).}

\item{...}{arguments to pass to methods.}
}
\description{
tbl_graph_to_treenetdf
}
