% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_explorer.R
\name{force_explorer}
\alias{force_explorer}
\title{Interactive \code{force_network} options explorer}
\usage{
force_explorer(data)
}
\arguments{
\item{data}{a network description in one of numerous forms (see details)}
}
\description{
An interactive shiny widget to explore the \code{force_network} options.
}
