% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_network.R
\name{force_network}
\alias{force_network}
\title{Create an interactive force network plot in a htmlwidget}
\usage{
force_network(data, width = NULL, height = NULL, ..., viewer = "internal")
}
\arguments{
\item{data}{a tree network description in one of numerous forms (see
details)}

\item{width, height}{width and height of exported htmlwidget in pixels (single integer value; default == NULL)}

\item{...}{other options (see details)}

\item{viewer}{whether to view the plot in the internal viewer or browser}
}
\description{
The \code{force_network} function creates an interactive force network plot in a
htmlwidget
}
