% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey_explorer.R
\name{sankey_explorer}
\alias{sankey_explorer}
\title{Interactive \code{sankey_network} options explorer}
\usage{
sankey_explorer(data)
}
\arguments{
\item{data}{a network description in one of numerous forms (see details)}
}
\description{
An interactive shiny widget to explore the \code{sankey_network} options.
}
