% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey_network.R
\name{sankey_network}
\alias{sankey_network}
\title{Create an interactive sankey network plot in a htmlwidget}
\usage{
sankey_network(data, width = NULL, height = NULL, ..., viewer = "browser")
}
\arguments{
\item{data}{a network description in one of numerous forms (see details)}

\item{width, height}{width and height of exported htmlwidget in pixels
(single integer value; default == NULL)}

\item{...}{other options (see details)}

\item{viewer}{whether to view the plot in the internal viewer or browser}
}
\description{
The \code{sankey_network} function creates an interactive sankey network plot in
a htmlwidget
}
