% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_png.R
\name{save_as_png}
\alias{save_as_png}
\title{Save a PNG screenshot of a htmlwidget}
\usage{
save_as_png(widget, filepath, background = "white", delay = 0.5)
}
\arguments{
\item{widget}{a htmlwidget}

\item{filepath}{a filepath where to save the file}

\item{background}{the background color underneath/behind the htmlwidget}

\item{delay}{a delay (in seconds) to wait before taking the screenshot}
}
\description{
The \code{save_as_png} function takes a screenshot of a htmlwidget as a PNG image.
}
