% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_svg.R
\name{save_as_svg}
\alias{save_as_svg}
\title{Save a SVG screenshot of a htmlwidget}
\usage{
save_as_svg(widget, filepath, background = "white", delay = 0.5)
}
\arguments{
\item{widget}{a htmlwidget}

\item{filepath}{a filepath where to save the file}

\item{background}{the background color underneath/behind the htmlwidget}

\item{delay}{a delay (in seconds) to wait before taking the screenshot}
}
\description{
The \code{save_as_svg} function takes a screenshot of a htmlwidget as a SVG image.
}
