% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_explorer.R
\name{tree_explorer}
\alias{tree_explorer}
\title{Interactive \code{tree_network} options explorer}
\usage{
tree_explorer(data)
}
\arguments{
\item{data}{a tree network description in one of numerous forms (see details)}
}
\description{
An interactive shiny widget to explore the \code{tree_network} options.
}
