% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_network.R
\name{tree_network}
\alias{tree_network}
\title{Create an interactive tree network plot in an htmlwidget}
\usage{
tree_network(
  data,
  width = NULL,
  height = NULL,
  treeType = "tidy",
  direction = "right",
  linkType = "diagonal",
  ...,
  viewer = "internal"
)
}
\arguments{
\item{data}{a tree network description in one of numerous forms (see
details)}

\item{width, height}{width and height of exported htmlwidget in pixels (single integer value; default == NULL)}

\item{treeType}{type of tree; one of "tidy" or "cluster" (see details) (default == "tidy")}

\item{direction}{direction toward which the tree grows; one of "right", "left", "down", or "up" (see details) (default == "right")}

\item{linkType}{type on link shape; one of "diagonal" or "elbow" (see details) (default == "diagonal")}

\item{...}{other options (see details)}

\item{viewer}{whether to view the plot in the internal viewer or browser}
}
\description{
The \code{tree_network} function creates an interactive tree network plot in an htmlwidget
}
\examples{
treedf <- data.frame(nodeId = LETTERS[1:7],
                     parentId = c("", "A", "A", "B", "B", "C", "C"),
                     name = LETTERS[1:7],
                     stringsAsFactors = FALSE)
tree_network(treedf)

}
